% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/power.R
\name{predictModelcustomer}
\alias{predictModelcustomer}
\title{This function reads the Neural Network model from the data directory and then predict the outcome based on the test data set.}
\usage{
predictModelcustomer()
}
\arguments{
\item{No}{input Required}
}
\description{
This function reads the Neural Network model from the data directory and then predict the outcome based on the test data set.
}
