% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/power.R
\name{readprojectpath}
\alias{readprojectpath}
\title{This function reads the project location from user and return the path to the caller.}
\usage{
readprojectpath()
}
\arguments{
\item{No}{input Required}
}
\description{
This function reads the project location from user and return the path to the caller.
}
