% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/power.R
\name{square}
\alias{square}
\title{The function returns a square of the given number}
\usage{
square(x)
}
\arguments{
\item{input}{a numeric value}
}
\description{
The function returns a square of the given number
}
