% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/power.R
\name{train_model}
\alias{train_model}
\title{The function train_model reads the data from the csv file, apply multiple models and save the models output as .rds files.}
\usage{
train_model()
}
\arguments{
\item{No}{input Required}
}
\description{
The function train_model reads the data from the csv file, apply multiple models and save the models output as .rds files.
}
